Added new filter autosuggest
